Make quality argument an int, do the same with MIN_QUALITY, MAX_QUALITY consts in Variables class

+++ core/src/com/gpteam/shopmanager/product/Product.java
@@ -77,7 +77,7 @@ public class Product {
         this.quality = quality;
     }
     
-    public void subQuality(short quality) {
+    public void subQuality(int quality) {
         this.quality -= quality;
     }
